requirements.txt
pandas~=1.4.2
tqdm~=4.64.0
sentencepiece
torch
transformers~=4.19.2
lizard~=1.17.10
langdetect~=1.0.9
langid~=1.1.6
pycld3
google-cloud-translate
regex~=2022.4.24
scikit-learn~=1.1.1
nltk~=3.7
numpy~=1.22.4
scipy~=1.8.1
python-weka-wrapper3~=0.2.9
xlrd~=1.2.0
imblearn~=0.0
javalang~=0.13.0